pyproject.toml
[project]
name = "invest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "lmkirvan", email = "kirvan.lewis@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.3",
    "polars>=1.36.1",
    "pyarrow>=22.0.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "typer>=0.20.0",
]

[project.scripts]
invest = "invest.main:app"
[build-system]
requires = ["uv_build>=0.8.3,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
